pyproject.toml
[tool.poetry]
authors = ["Tiago Requeijo <tiago.requeijo.dev@gmail.com>"]
description = "A library to load configuration parameters from multiple sources and formats"
homepage = "https://github.com/tr11/python-configuration"
include = ["config/py.typed"]
keywords = ['configuration', 'settings', 'json', 'yaml', 'toml', 'ini']
license = "MIT"
name = "python-configuration"
packages = [{ include = "config" }]
readme = 'README.md'
repository = "https://github.com/tr11/python-configuration"
version = "0.8.3"

[tool.poetry.dependencies]
azure-identity = { version = "^1.5.0", optional = true }
azure-keyvault = { version = "^4.0.0", optional = true }
boto3 = { version = "^1.10.48", optional = true }
google-cloud-secret-manager = { version = "^2.0.0", optional = true }
python = "^3.6.2"
pyyaml = { version = "^6.0", optional = true }
toml = { version = "^0.10.0", optional = true }

[tool.poetry.dev-dependencies]
flake8-blind-except = "^0.2.0"
flake8-bugbear = "^21.11"
flake8-builtins = "^1.4"
flake8-comprehensions = "^3.3.1"
flake8-docstrings = "^1.3"
flake8-import-order = "^0.18.0"
mypy = "^0.921"
pydocstyle = "^6.0"
pytest = "^6.2.2"
pytest-black = "^0.3.12"
pytest-cov = "^3.0"
pytest-flake8 = "^1.0"
pytest-mypy = "^0.8.0"
pytest-mock = "^3.5.0"
sphinx = "^4.3"
sphinx-autodoc-typehints = "^1.11.0"
# # need to allow pre-releases to address a recent change in pytest-black. See https://github.com/shopkeep/pytest-black/issues/45
black = { version = "21.12b0*", allow-prereleases = true }

[tool.poetry.extras]
aws = ["boto3"]
azure = ["azure-keyvault", "azure-identity"]
gcp = ["google-cloud-secret-manager"]
toml = ["toml"]
yaml = ["pyyaml"]

[tool.black]
line-length = 88

[tool.tox]
legacy_tox_ini = """
[tox]
isolated_build = true
envlist = py36, py37, py38, py39, py310

[testenv]
whitelist_externals = poetry
commands =
    poetry install -v -E toml -E yaml -E azure -E aws -E gcp
    poetry run pytest
"""

[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=1.1.4"]
